Factor out key path instantiation to reuse in the compatibility library.
The Swift 5.0 and 5.1 layouts for key path objects are identical, so we can use the same
instantiation logic for both, and bug fixes and/or extensions we make to the 5.1 runtime can be
back-deployed to 5.0 using the compatibility override for swift_getKeyPath.
